Add amd64 builds; Switch to using bridged network
Due to current development being done on amd64 we have new Dockerfiles
to support this.

A few improvements which will eventually be carried over to ARM:

- base image has been bumped to the latest stable debian
- base images for both prometheus and grafana are updated to latest
- base container is upgraded on build, getting any security patches
- we switch to PID 1 being dumb-init, which ensures we handle any
  incoming process signals correctly[1]

Switched to using docker bridged networks rather than host networks.  I
needed this in particular because I have enabled user-namespace
separation.  This adds security, but prevents containers from binding
to the host network.

The only down side is we need the IP address of the LightwaveRF Link
station configured before we run the app - i.e. we can't depend on
sending broadcast packets as they won't traverse the bridge.  The
configuration file format is therefore updated to add support for
unicast communication with the LightwaveRF Link station.

On the up size, with the switch to bridged networking we're now able
to reference the parallel containers by their name in the docker-compose file,
rather than just by their port.

1: https://engineeringblog.yelp.com/2016/01/dumb-init-an-init-for-docker.html