Tweak Grafana dashboard
I found the stale chart and alerts weren't working for me as my system
was failing to receive updates from several devices, but the alert
didn't trigger, and it was hard to see what was going on in the graph.

As a result I've set the graph to log-2 to make differences more visible,
and set the alert threshold much lower and to alert when there's no
data.