Call for heat when total valve ouput above threshold
The boiler pump struggles when there isn't enough output enabled across
the valves.  To avoid this we now only turn on the boiler if there's at
least 2 valves open.  Given the valves open in 10% increments we just
set this threshold to be above 10%.  Lets see if it resolves the boiler
issue without causing responsiveness to be negatively impacted.

Remove Prometheus and Grafana containers from the composer config as
we're using our centralized stack.

Add handling for detecting when the Link has been rebooted.  This was
needed because the app would get into a bad state.